Old version had the following issues:
(Issue 1)
- format is BigTiff
- XResolution/YResolution (Single float value) has been written
- When trying to load the following situation happens:

In BigTiff format all values up to 64 bit can be directly written into the
offset, the old implementation only handleded those situations up to 32 bit
(forgot to port that bit)

Issue is now fixed.

(Issue 2)

Transforming from normal tiff to bigtiff mid-file (MakeBigTiff()) was not complete
some parts of it were missing. Some logic in writing 33 to 64 bits directly into the
offset in bigtiff-format were also missing.

Problem: objects between 33 and 64 bits had to be read from the offset location
and then be rewritten directly into the offset (because offset grows from 32 to 64
bit in bigtiff)

Issue is now fixed.

All tests run without problems.

+++ LibTiff/Tiff.cs
@@ -6184,13 +6184,14 @@ namespace BitMiracle.LibTiff.Classic
         }
         private static long[] IntToLong(int[] inputArray)
         {
-            long[] output = new long[inputArray.Length];
+          long[] output = new long[inputArray.Length];
           for (int i = 0; i < inputArray.Length; i++)
           {
-            output[i] = (long)inputArray[i];
+            output[i] = (uint)inputArray[i];
           }
-            return output;
+          return output;
         }
+
         private static uint[] LongToInt(ulong[] inputArray)
         {
           uint[] output = new uint[inputArray.Length];
